fix implicit casting

+++ Code/Engine/Physics/Collision2D.cpp
@@ -24,12 +24,12 @@ bool _SetManifold(Manifold2D& out_manifold, const AABB2& a, const AABB2& b)
 		float height = manifold.GetHeight();
 		Vec2 displacement = a.GetCenter() - b.GetCenter();
 		if (width < height) {
-			float dir = Sgn(Vec2(1, 0).DotProduct(displacement));
-			out_manifold.normal = Vec2(dir, 0);
+			float dir = (float)Sgn(Vec2(1.f, 0.f).DotProduct(displacement));
+			out_manifold.normal = Vec2(dir, 0.f);
 			out_manifold.penetration = width;
 		} else {
-			float dir = Sgn(Vec2(0, 1).DotProduct(displacement));
-			out_manifold.normal = Vec2(0, dir);
+			float dir = (float)Sgn(Vec2(0.f, 1.f).DotProduct(displacement));
+			out_manifold.normal = Vec2(0.f, dir);
 			out_manifold.penetration = height;
 		}
 		return true;
